ml_api/requirements-minimal.txt
Flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
numpy>=1.26.0
scikit-learn>=1.3.2
joblib>=1.3.2
pandas>=2.1.4
python-dotenv==1.0.0

